[SERVER_DEV78] Fix "Some Issue" // DEV_78
 - UserNode에 Index를 BYTE로 변환하여 갖고 있도록 변경함

수정 이슈
1. 파일 입출력 시에는, 관리를 위해 int로 처리하고 바로 Byte로 형변환 함
2. 상점에서 구매하면 돈이 두배로 드는 이슈
3. 구매 성공, 실패 UI가 둘다 출력되는 이슈
4. 로그인 한 번한 계정이 다시 로그인할 때, 서버 터지는 이슈
5. 구매한 캐릭터가 변경되는 이슈
7. 구매 요청 시, 요청이 2번 가는 이슈

+++ Server/Server/SceneNetworkManager/_1_LoginScene/LoginScene.cpp
@@ -36,7 +36,7 @@ void SCENE_NETWORK_MANAGER::LoginScene::_RecvDemandLogin(SocketInfo* pClient)
 	int outTitleBit{ 0 };
 	int outCharacterBit{ 0 };
 	vector<Type_Nickname> outFriendStringCont;
-	int outActiveCharacterBit{ 0 };
+	BYTE outActiveCharacterBit{ 0 };
 
 	//typeBuffer == 1 
 	//? LoginTest(pClient, InUserData, idBuffer, pwBuffer, outWinCount, outLoseCount, outMoney) 
@@ -47,8 +47,9 @@ void SCENE_NETWORK_MANAGER::LoginScene::_RecvDemandLogin(SocketInfo* pClient)
 		; failReason)
 		__SendFailLogin(pClient, failReason);
 	else 
-		__SendPermitLogin(pClient, outNickname, outWinCount, outLoseCount, outMoney, outAchievementBit, outTitleBit, outCharacterBit, outActiveCharacterBit, outFriendStringCont);
-		std::cout << "     [UserDataManager] " << outNickname << " (" << idBuffer << ")" <<" ´ÔÀÌ ·Î±×ÀÎ ÇÏ¼Ì½À´Ï´Ù. " << "\n";
+		__SendPermitLogin(pClient, outNickname, outWinCount, outLoseCount, outMoney, outAchievementBit, outTitleBit, outCharacterBit, static_cast<int>(outActiveCharacterBit), outFriendStringCont);
+	
+	std::cout << "     [UserDataManager] " << outNickname << " (" << idBuffer << ")" <<" ´ÔÀÌ ·Î±×ÀÎ ÇÏ¼Ì½À´Ï´Ù. " << "\n";
 }
 
 void SCENE_NETWORK_MANAGER::LoginScene::__SendFailLogin(SocketInfo* pClient, const int RetFailReason)
